changed incrementPeopleProperties to double

+++ lib/native_mixpanel.dart
@@ -76,7 +76,7 @@ class Mixpanel extends _Mixpanel {
     return this._mp.track('setPeopleProperties', jsonEncode(props));
   }
 
-  Future incrementPeopleProperties(Map<String, int> props) {
+  Future incrementPeopleProperties(Map<String, double> props) {
     return this._mp.track('incrementPeopleProperties', jsonEncode(props));
   }
